add back in create function hook and restore tests
