+++ pg_analytics/src/federation/executor.rs
@@ -41,7 +41,7 @@ impl SQLExecutor for ColumnExecutor {
         sql: &str,
         schema: SchemaRef,
     ) -> Result<SendableRecordBatchStream, DataFusionError> {
-        let logical_plan = LogicalPlan::try_from(QueryString(sql))?;
+        let (logical_plan, _) = <(LogicalPlan, bool)>::try_from(QueryString(sql))?;
         let batch_stream = Session::with_session_context(|context| {
             Box::pin(async move {
                 let dataframe = context.execute_logical_plan(logical_plan.clone()).await?;
