+++ pg_analytics/src/hooks/mod.rs
@@ -1,4 +1,5 @@
 mod alter;
+mod createfunction;
 mod drop;
 mod executor;
 mod handler;
