+++ pg_analytics/tests/udf.rs
@@ -5,7 +5,7 @@ use pretty_assertions::assert_eq;
 use rstest::*;
 use sqlx::PgConnection;
 
-#[rstest]
+#[rs_test]
 fn udf(mut conn: PgConnection) {
     r#"
         CREATE TABLE x (
@@ -72,20 +72,23 @@ fn udf_deletion(mut conn: PgConnection) {
     .execute(&mut conn);
 
     // This is the current behavior, but we want deletion to actually work in the future!
-    let rows: Vec<(i32,)> = "SELECT add(a, b) FROM x".fetch(&mut conn);
-    let sums: Vec<i32> = rows.into_iter().map(|r| r.0).collect();
-    assert_eq!(sums, [3, 7, 11, 15]);
+    match "SELECT add(a, b) FROM x".execute_result(&mut conn) {
+        Err(err) => assert!(err
+            .to_string()
+            .contains("No function matches the given name")),
+        _ => panic!("Deleted functions should not execute"),
+    };
 }
 
 #[rstest]
 fn udf_coercion(mut conn: PgConnection) {
     r#"
         CREATE TABLE x (
-            a int8,
-            b int8
+            a INTEGER,
+            b INTEGER
         ) USING parquet;
         INSERT INTO x VALUES (1, 2), (3, 4), (5, 6), (7, 8);
-        CREATE FUNCTION add(int8, int8) RETURNS int8
+        CREATE FUNCTION add(integer, integer) RETURNS integer
             AS 'select $1 + $2;'
             LANGUAGE SQL
             IMMUTABLE
@@ -93,7 +96,12 @@ fn udf_coercion(mut conn: PgConnection) {
     "#
     .execute(&mut conn);
 
-    let rows: Vec<(i32,)> = "SELECT add(1::int4, b) FROM x".fetch(&mut conn);
+    match "SELECT add(1, b) FROM x".execute_result(&mut conn) {
+        Err(err) => assert!(err.to_string().contains("expected Int64 but found Int32")),
+        _ => panic!("Constant coercion should not work"),
+    };
+
+    let rows: Vec<(i32,)> = "SELECT add(1::integer, b) FROM x".fetch(&mut conn);
     let sums: Vec<i32> = rows.into_iter().map(|r| r.0).collect();
     assert_eq!(sums, [3, 5, 7, 9]);
 }
